fix up ldp, move Context, Proof and URI into the crate
