+++ ssi-ldp/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "ssi-ldp"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0"
+async-trait = "0.1"
+lazy_static = "1.4"
+bs58 = { version = "0.4", features = ["check"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+serde_jcs = "0.1"
+hex = "0.4"
+multibase = "0.8"
+ssi-jwk = { path = "../ssi-jwk" }
+ssi-json-ld = { path = "../ssi-json-ld" }
+ssi-core = { path = "../ssi-core" }
+ssi-dids = { path = "../ssi-dids" }
+ssi-crypto = { path = "../ssi-crypto" }
+ssi-jws = { path = "../ssi-jws" }
+ssi-jwt = { path = "../ssi-jwt" }
+ssi-tzkey = { path = "../ssi-tzkey" }
+caips = { path = "../caips" }
+ssi-contexts = { version = "0.1.3", path = "../contexts" }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+chrono = { version = "0.4", features = ["serde", "wasmbind"] }
